Add a rudimentary whitespace stripper (minifier) for HTML.
Please do not use this builder in production environments
as it does not actually minify HTML and simply
trims each line and removes EOL characters from the source.
This is dangerous practice if you don't know what you are
doing.  Currently there is no parser to ignore inline scripts.
In the future a parser that delegates inline script processing
to an external minifier will be added so this becomes safer.